[API] correções no Model e Controller de Teams

+++ api/controllers/teams.py
@@ -24,34 +24,45 @@ class TeamsController(Controller):
             return Controller.format_response(status_code=404)
             
     @Controller.route('/teams', methods=['POST'])
-    def create():
+    @Controller.authenticate_user
+    def create(): #TESTAR Adicionar owner_login
         data = request.get_json()
-        team = TeamsModel(data['initials'], data['name'])
+        current_user = Controller.authenticated_user()
+        team = TeamsModel(data['initials'], data['name'],current_user)
         team.insert()
         return Controller.format_response(status_code=201)
 
     @Controller.route('/teams/<initials>', methods=['PUT'])
-    def update(initials):
+    @Controller.authenticate_user
+    def update(initials): #TESTAR averiguar se o usuário autenticado é o dono do time
         team = TeamsModel.find_by_initials(initials)
         if team: #Verifica se possui conteúdo
             team = team[0]
             data = request.get_json()
-            if 'name' in data:
-                team.name = data['name']
-                team.update()
-                return Controller.format_response(status_code=200)
+            current_user = Controller.authenticated_user()
+            
+            if not current_user.login == team.owner_login:
+                return Controller.format_response(errors=13, status_code=403) 
 
-            if 'initials' in data:
-                team.initials = data['initials']
-                team.update()
-                return Controller.format_response(status_code=200)
+            for field in ('initials', 'name','owner_login'):
+                if field in data:
+                    setattr(team, field, data[field])
+            team.update()
+            return Controller.format_response(status_code=200)
+
         else:
             return Controller.format_response(status_code=404)
 
     
     @Controller.route('/teams/<initials>', methods=['DELETE'])
-    def delete(initials):
+    @Controller.authenticate_user
+    def delete(initials): #TESTAR averiguar se o usuário autenticado é o dono do time
         team = TeamsModel.find_by_initials(initials)
+        current_user = Controller.authenticated_user()
+        
+        if not current_user.login == team.owner_login:
+            return Controller.format_response(errors=13, status_code=403) 
+
         if team:
             team = team[0]
             team.delete()
